+++ include/Ark/Compiler/Compiler.hpp
@@ -27,7 +27,7 @@
 #include <Ark/Compiler/BytecodeReader.hpp>
 #include <Ark/Builtins/Builtins.hpp>
 #include <Ark/Utils.hpp>
-#include <Ark/Config.hpp>
+#include <Ark/Platform.hpp>
 #include <Ark/Compiler/makeErrorCtx.hpp>
 
 namespace Ark