+++ include/Ark/VM/UserType.hpp
@@ -16,7 +16,7 @@
 #include <vector>
 #include <utility>
 
-#include <Ark/Config.hpp>
+#include <Ark/Platform.hpp>
 
 namespace Ark
 {